+++ src/OpenTelemetry/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Added `AddProcessor` overload which exposes the factory pattern
+  `(Func<IServiceProvider, BaseProcessor<LogRecord>> implementationFactory)`.
+  ([#4916](https://github.com/open-telemetry/opentelemetry-dotnet/pull/4916))
+
 * Updated `Microsoft.Extensions.Logging` package version to
   `8.0.0-rc.2.23479.6`.
   ([#4959](https://github.com/open-telemetry/opentelemetry-dotnet/pull/4959))